Fix issue with HTTP calls if protocol not specified. Restructure HTTP util functions
